btcservice/requirements.txt
falcon
gunicorn
pycrypto